% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rohirrim.R
\name{rohirrim}
\alias{rohirrim}
\title{Calculation of Over-prediction rate, Under-prediction rate, Potential Presence Increment, and Potential Aabsence Increment}
\usage{
rohirrim(obj)
}
\arguments{
\item{confMatrix}{An object of class data.frame created by the confMatrix function from the SDMtune package.}
}
\value{
A data.frame with the OPR, UPR, PPI, and PAI values.
}
\description{
This function calculates the OPR, UPR, PPI, and PAI values from the confMatrix function of the SDMtune package.
}
